refactor(parquet-datasource): mechanical cleanup — split opener.rs, file_format.rs, row_group_filter.rs
A series of pure code-motion splits to make the parquet datasource crate
easier to navigate and modify. No public API changes (existing import
paths preserved via re-exports), no behavior changes.

Five separate moves:

1. `opener.rs` (~2,700 LOC) → `opener/` module:
   - `opener/early_stop.rs` — `EarlyStoppingStream`, the dynamic-filter
     early-termination wrapper used at the end of `build_stream`.
   - `opener/encryption.rs` — `EncryptionContext` + the
     `ParquetMorselizer::get_encryption_context` helpers. Isolates the
     `#[cfg(feature = "parquet_encryption")]` gating.
   - `opener/push_decoder_stream.rs` — `PushDecoderStreamState`, the
     inner stream state machine that drives a `ParquetPushDecoder`.

2. `file_format.rs` (~2,000 LOC) split:
   - `sink.rs` — `ParquetSink` + the parallel-write machinery
     (`column_serializer_task`, `spawn_column_parallel_row_group_writer`,
     `output_single_parquet_file_parallelized`, etc.). The historical
     `file_format::ParquetSink` path is preserved via `pub use`.
   - `schema_coercion.rs` — Arrow-schema coercion utilities
     (`apply_file_schema_type_coercions`, `coerce_int96_to_resolution`,
     `coerce_file_schema_to_view_type`, `coerce_file_schema_to_string_type`,
     `transform_schema_to_view`, `transform_binary_to_string`,
     `field_with_new_type`) and their tests. Re-exported at the crate
     root for backward compat.

3. `row_group_filter.rs` (~1,900 LOC):
   - `bloom_filter.rs` — `BloomFilterStatistics` (the loaded SBBF data
     + its `PruningStatistics` adapter). Separates "data we loaded
     from the file" from "the access-plan filter that consumes it".

After this PR:
- `opener.rs`: 2,717 → 2,433 LOC (-10%; further drop possible by
  extracting the test module, but tests stay near the code under
  test for now).
- `file_format.rs`: 2,038 → 633 LOC (-69%).
- `row_group_filter.rs`: 1,929 → 1,756 LOC (-9%; bloom code moved
  out, the file is now focused entirely on `RowGroupAccessPlanFilter`).
